allow deletion of duplicate relations
A duplicate relation shouldn't happen. But if it does, it should be
deletable.

Previous code would use the `.id` of the relation to compare them.
However, the .id is NOT the id of the relation, but the id of the
object being related to. So we should compare using the `.relationId`